+++ apis/s3/src/main/java/org/jclouds/s3/binders/BindPartIdsAndETagsToRequest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jclouds.aws.s3.binders;
+package org.jclouds.s3.binders;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
