+++ apis/s3/src/main/java/org/jclouds/s3/blobstore/S3BlobStore.java
@@ -49,6 +49,7 @@ import org.jclouds.s3.blobstore.functions.BucketToResourceList;
 import org.jclouds.s3.blobstore.functions.ContainerToBucketListOptions;
 import org.jclouds.s3.blobstore.functions.ObjectToBlob;
 import org.jclouds.s3.blobstore.functions.ObjectToBlobMetadata;
+import org.jclouds.s3.blobstore.strategy.MultipartUploadStrategy;
 import org.jclouds.s3.domain.AccessControlList;
 import org.jclouds.s3.domain.AccessControlList.GroupGranteeURI;
 import org.jclouds.s3.domain.AccessControlList.Permission;
@@ -77,6 +78,7 @@ public class S3BlobStore extends BaseBlobStore {
    private final BlobToHttpGetOptions blob2ObjectGetOptions;
    private final Provider<FetchBlobMetadata> fetchBlobMetadataProvider;
    private final LoadingCache<String, AccessControlList> bucketAcls;
+   protected final Provider<MultipartUploadStrategy> multipartUploadStrategy;
 
    @Inject
    protected S3BlobStore(BlobStoreContext context, BlobUtils blobUtils, Supplier<Location> defaultLocation,
@@ -85,7 +87,8 @@ public class S3BlobStore extends BaseBlobStore {
             ContainerToBucketListOptions container2BucketListOptions, BucketToResourceList bucket2ResourceList,
             ObjectToBlob object2Blob, BlobToHttpGetOptions blob2ObjectGetOptions, BlobToObject blob2Object,
             ObjectToBlobMetadata object2BlobMd, Provider<FetchBlobMetadata> fetchBlobMetadataProvider,
-            LoadingCache<String, AccessControlList> bucketAcls) {
+            LoadingCache<String, AccessControlList> bucketAcls,
+            Provider<MultipartUploadStrategy> multipartUploadStrategy) {
       super(context, blobUtils, defaultLocation, locations);
       this.blob2ObjectGetOptions = checkNotNull(blob2ObjectGetOptions, "blob2ObjectGetOptions");
       this.sync = checkNotNull(sync, "sync");
@@ -97,6 +100,7 @@ public class S3BlobStore extends BaseBlobStore {
       this.object2BlobMd = checkNotNull(object2BlobMd, "object2BlobMd");
       this.fetchBlobMetadataProvider = checkNotNull(fetchBlobMetadataProvider, "fetchBlobMetadataProvider");
       this.bucketAcls = checkNotNull(bucketAcls, "bucketAcls");
+      this.multipartUploadStrategy = checkNotNull(multipartUploadStrategy, "multipartUploadStrategy");
    }
 
    /**
@@ -249,6 +253,11 @@ public class S3BlobStore extends BaseBlobStore {
     */
    @Override
    public String putBlob(String container, Blob blob, PutOptions overrides) {
+      if (overrides.isMultipart()) {
+         // need to use a provider if the strategy object is stateful
+         return multipartUploadStrategy.get().execute(container, blob);
+      }
+
       // TODO: Make use of options overrides
       PutObjectOptions options = new PutObjectOptions();
       try {
