+++ apis/s3/src/main/java/org/jclouds/s3/blobstore/strategy/AsyncMultipartUploadStrategy.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jclouds.aws.s3.blobstore.strategy;
+package org.jclouds.s3.blobstore.strategy;
 
-import org.jclouds.aws.s3.blobstore.strategy.internal.ParallelMultipartUploadStrategy;
 import org.jclouds.blobstore.domain.Blob;
 import org.jclouds.blobstore.options.PutOptions;
+import org.jclouds.s3.blobstore.strategy.internal.ParallelMultipartUploadStrategy;
 
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.inject.ImplementedBy;
