+++ apis/s3/src/main/java/org/jclouds/s3/blobstore/strategy/MultipartUpload.java
@@ -14,19 +14,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jclouds.aws.s3.functions;
+package org.jclouds.s3.blobstore.strategy;
 
-import javax.inject.Singleton;
+public final class MultipartUpload {
 
-import org.jclouds.s3.domain.ObjectMetadata;
+   /* Maximum number of parts per upload */
+   public static final int MAX_NUMBER_OF_PARTS = 10000;
+   /* Maximum number of parts returned for a list parts request */
+   public static final int MAX_LIST_PARTS_RETURNED = 1000;
+   /* Maximum number of multipart uploads returned in a list multipart uploads request */
+   public static final int MAX_LIST_MPU_RETURNED = 1000;
 
-import com.google.common.base.Function;
+   /*
+    * part size 5 MB to 5 GB, last part can be < 5 MB
+    */
+   public static final long MIN_PART_SIZE = 5242880L;
+   public static final long MAX_PART_SIZE = 5368709120L;
 
-@Singleton
-public class ObjectMetadataKey implements Function<Object, String> {
-
-   public String apply(Object from) {
-      return ((ObjectMetadata) from).getKey();
+   private MultipartUpload() {
+      throw new AssertionError("intentionally unimplemented");
    }
-
 }
