+++ apis/s3/src/test/java/org/jclouds/s3/blobstore/strategy/internal/MpuGraphData.java
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jclouds.aws.s3.blobstore.strategy.internal;
+package org.jclouds.s3.blobstore.strategy.internal;
 
-import org.jclouds.aws.s3.blobstore.strategy.MultipartUpload;
+import org.jclouds.s3.blobstore.strategy.MultipartUpload;
 
 /**
  * Print out on the console some graph data regarding the partitioning algorithm.
