+++ apis/s3/src/test/resources/initiate-multipart-upload.xml
