+++ providers/aws-s3/src/test/java/org/jclouds/aws/s3/AWSS3ClientTest.java
@@ -20,31 +20,21 @@ import static org.jclouds.reflect.Reflection2.method;
 import static org.testng.Assert.assertEquals;
 
 import java.io.IOException;
-import java.util.Map;
 import java.util.Set;
 
-import org.jclouds.Fallbacks.VoidOnNotFoundOr404;
 import org.jclouds.aws.s3.config.AWSS3HttpApiModule;
 import org.jclouds.aws.s3.filters.AWSRequestAuthorizeSignature;
-import org.jclouds.aws.s3.functions.ETagFromHttpResponseViaRegex;
-import org.jclouds.aws.s3.functions.UploadIdFromHttpResponseViaRegex;
 import org.jclouds.blobstore.binders.BindBlobToMultipartFormTest;
 import org.jclouds.date.TimeStamp;
-import org.jclouds.fallbacks.MapHttp4xxCodesToExceptions;
 import org.jclouds.http.HttpRequest;
 import org.jclouds.http.functions.ParseETagHeader;
 import org.jclouds.http.functions.ParseSax;
-import org.jclouds.http.functions.ReleasePayloadAndReturn;
 import org.jclouds.http.functions.ReturnTrueIf2xx;
-import org.jclouds.io.Payload;
-import org.jclouds.io.Payloads;
 import org.jclouds.location.Region;
 import org.jclouds.rest.ConfiguresHttpApi;
 import org.jclouds.rest.internal.GeneratedHttpRequest;
 import org.jclouds.s3.S3Client;
 import org.jclouds.s3.S3ClientTest;
-import org.jclouds.s3.domain.ObjectMetadata;
-import org.jclouds.s3.domain.ObjectMetadataBuilder;
 import org.jclouds.s3.domain.S3Object;
 import org.jclouds.s3.fallbacks.FalseIfBucketAlreadyOwnedByYouOrOperationAbortedWhenBucketExists;
 import org.jclouds.s3.options.CopyObjectOptions;
@@ -157,95 +147,6 @@ public class AWSS3ClientTest extends S3ClientTest<AWSS3Client> {
       assertResponseParserClassEquals(method, request, ReturnTrueIf2xx.class);
       assertSaxResponseParserClassEquals(method, null);
       assertFallbackClassEquals(method, FalseIfBucketAlreadyOwnedByYouOrOperationAbortedWhenBucketExists.class);
-
-      checkFilters(request);
-   }
-
-   public void testInitiateMultipartUpload() throws SecurityException, NegativeArraySizeException,
-         NoSuchMethodException {
-      Invokable<?, ?> method = method(AWSS3Client.class, "initiateMultipartUpload", String.class, ObjectMetadata.class,
-            PutObjectOptions[].class);
-      GeneratedHttpRequest request = processor.createRequest(method, ImmutableList.<Object> of("bucket", ObjectMetadataBuilder.create().key("foo")
-            .contentMD5(new byte[16]).build()));
-
-      assertRequestLineEquals(request, "POST https://bucket." + url + "/foo?uploads HTTP/1.1");
-      assertNonPayloadHeadersEqual(request,
-            "Content-MD5: AAAAAAAAAAAAAAAAAAAAAA==\n" +
-            "Content-Type: binary/octet-stream\n" +
-            "Host: bucket." + url + "\n");
-      assertPayloadEquals(request, null, null, false);
-
-      // as this is a payload-related command, but with no payload, be careful
-      // that we check
-      // filtering and do not ignore if this fails later.
-      request = (GeneratedHttpRequest) request.getFilters().get(0).filter(request);
-
-      assertRequestLineEquals(request, "POST https://bucket." + url + "/foo?uploads HTTP/1.1");
-      assertNonPayloadHeadersEqual(request,
-            "Authorization: AWS identity:972m/Bqn2L5FIaB+wWDeY83mGvU=\n" +
-            "Content-MD5: AAAAAAAAAAAAAAAAAAAAAA==\n" +
-            "Content-Type: binary/octet-stream\n" +
-            "Date: 2009-11-08T15:54:08.897Z\n" +
-            "Host: bucket." + url + "\n");
-      assertPayloadEquals(request, null, null, false);
-
-      assertResponseParserClassEquals(method, request, UploadIdFromHttpResponseViaRegex.class);
-      assertSaxResponseParserClassEquals(method, null);
-      assertFallbackClassEquals(method, MapHttp4xxCodesToExceptions.class);
-
-      checkFilters(request);
-   }
-
-   public void testAbortMultipartUpload() throws SecurityException, NegativeArraySizeException, NoSuchMethodException {
-      Invokable<?, ?> method = method(AWSS3Client.class, "abortMultipartUpload", String.class, String.class, String.class);
-      GeneratedHttpRequest request = processor.createRequest(method, ImmutableList.<Object> of("bucket", "foo", "asdsadasdas", 1,
-            Payloads.newStringPayload("")));
-
-      assertRequestLineEquals(request, "DELETE https://bucket." + url + "/foo?uploadId=asdsadasdas HTTP/1.1");
-      assertNonPayloadHeadersEqual(request, "Host: bucket." + url + "\n");
-      assertPayloadEquals(request, "", "application/unknown", false);
-
-      assertResponseParserClassEquals(method, request, ReleasePayloadAndReturn.class);
-      assertSaxResponseParserClassEquals(method, null);
-      assertFallbackClassEquals(method, VoidOnNotFoundOr404.class);
-
-      checkFilters(request);
-   }
-
-   public void testUploadPart() throws SecurityException, NegativeArraySizeException, NoSuchMethodException {
-      Invokable<?, ?> method = method(AWSS3Client.class, "uploadPart", String.class, String.class, int.class,
-            String.class, Payload.class);
-      GeneratedHttpRequest request = processor.createRequest(method, ImmutableList.<Object> of("bucket", "foo", 1, "asdsadasdas",
-            Payloads.newStringPayload("")));
-
-      assertRequestLineEquals(request, "PUT https://bucket." + url + "/foo?partNumber=1&uploadId=asdsadasdas HTTP/1.1");
-      assertNonPayloadHeadersEqual(request, "Host: bucket." + url + "\n");
-      assertPayloadEquals(request, "", "application/unknown", false);
-
-      assertResponseParserClassEquals(method, request, ParseETagHeader.class);
-      assertSaxResponseParserClassEquals(method, null);
-      assertFallbackClassEquals(method, MapHttp4xxCodesToExceptions.class);
-
-      checkFilters(request);
-   }
-
-   public void testCompleteMultipartUpload() throws SecurityException, NegativeArraySizeException,
-         NoSuchMethodException {
-      Invokable<?, ?> method = method(AWSS3Client.class, "completeMultipartUpload", String.class, String.class,
-            String.class, Map.class);
-      GeneratedHttpRequest request = processor.createRequest(method, ImmutableList.<Object> of("bucket", "foo", "asdsadasdas",
-            ImmutableMap.<Integer, String> of(1, "\"a54357aff0632cce46d942af68356b38\"")));
-
-      assertRequestLineEquals(request, "POST https://bucket." + url + "/foo?uploadId=asdsadasdas HTTP/1.1");
-      assertNonPayloadHeadersEqual(request, "Host: bucket." + url + "\n");
-      assertPayloadEquals(
-            request,
-            "<CompleteMultipartUpload><Part><PartNumber>1</PartNumber><ETag>\"a54357aff0632cce46d942af68356b38\"</ETag></Part></CompleteMultipartUpload>",
-            "text/xml", false);
-
-      assertResponseParserClassEquals(method, request, ETagFromHttpResponseViaRegex.class);
-      assertSaxResponseParserClassEquals(method, null);
-      assertFallbackClassEquals(method, MapHttp4xxCodesToExceptions.class);
 
       checkFilters(request);
    }
